feat: created framework for favourites; current framework creates a query based on each id from the database and logs the name of that movie; TODO: extend query to create a card result from the query
