+++ src/components/myMoviesTab/TabContent.tsx
@@ -7,13 +7,21 @@ const TabContent = () => {
     // add the result of these queries to a result array map and then parse that 
     // parse this through via a list of cards and display under the tab-main-content
 
+    collectMovieIdDatabase("favourites")
+
+    console.log("----------")
+
+    collectMovieIdDatabase("seen")
+
     collectMovieIdDatabase("watchlist")
 
-  return (
-    <div className="tab-main-content">
-        <h1>Placeholder for movie cards</h1>
-    </div>
-  )
+
+
+    return (
+        <div className="tab-main-content">
+            <h1>Placeholder for movie cards</h1>
+        </div>
+    )
 }
 
 export default TabContent
